+++ components/evaluation/README.md
@@ -2,5 +2,8 @@
 
 This directory contains components in the **Evaluation** category:
 
-- [Evalhub Kserve](./evalhub_kserve/README.md): Evaluate a model via Eval Hub with a KServe InferenceService.
 - [Lm Eval](./lm_eval/README.md): A Universal LLM Evaluator component using EleutherAI's lm-evaluation-harness.
+
+## Subcategories
+
+- [Evalhub](./evalhub/README.md)
