+++ components/evaluation/evalhub/README.md
@@ -1,0 +1,5 @@
+# Evalhub
+
+This subcategory contains components in the **Evalhub** group:
+
+- [Evalhub Kserve](./kserve/README.md): Evaluate a model via Eval Hub with a KServe InferenceService.
