+++ components/evaluation/evalhub/__init__.py
@@ -1,0 +1,4 @@
+"""Eval Hub Components Subcategory.
+
+Components for evaluating models via the Eval Hub service.
+"""
