+++ pipelines/training/README.md
@@ -7,3 +7,4 @@ This directory contains pipelines in the **Training** category:
 - [Automl](./automl/README.md)
 - [Autorag](./autorag/README.md)
 - [Finetuning](./finetuning/README.md)
+- [Finetuning Evalhub](./finetuning_evalhub/README.md)
