+++ pipelines/training/finetuning/README.md
@@ -2,15 +2,9 @@
 
 This subcategory contains pipelines in the **Finetuning** group:
 
-- [Lora Evalhub Pipeline](./lora_evalhub/README.md): LoRA Training Pipeline with Eval Hub evaluation (KServe).
-- [Lora Minimal Evalhub Pipeline](./lora_minimal_evalhub/README.md): LoRA Training Pipeline — Eval Hub Minimal.
 - [Lora Minimal Pipeline](./lora_minimal/README.md): LoRA Minimal Training Pipeline - Parameter-efficient fine-tuning.
 - [Lora Pipeline](./lora/README.md): LoRA Training Pipeline - Parameter-efficient fine-tuning.
-- [Osft Evalhub Pipeline](./osft_evalhub/README.md): OSFT Training Pipeline with Eval Hub evaluation (KServe).
-- [Osft Minimal Evalhub Pipeline](./osft_minimal_evalhub/README.md): OSFT Training Pipeline — Eval Hub Minimal.
 - [Osft Minimal Pipeline](./osft_minimal/README.md): OSFT Minimal Training Pipeline - Continual learning without catastrophic forgetting.
 - [Osft Pipeline](./osft/README.md): OSFT Training Pipeline - Continual learning without catastrophic forgetting.
-- [Sft Evalhub Pipeline](./sft_evalhub/README.md): SFT Training Pipeline with Eval Hub evaluation (KServe).
-- [Sft Minimal Evalhub Pipeline](./sft_minimal_evalhub/README.md): SFT Training Pipeline — Eval Hub Minimal.
 - [Sft Minimal Pipeline](./sft_minimal/README.md): SFT Training Pipeline - Standard supervised fine-tuning with instructlab-training.
 - [Sft Pipeline](./sft/README.md): SFT Training Pipeline - Standard supervised fine-tuning with instructlab-training.
