+++ pipelines/training/finetuning_evalhub/README.md
@@ -1,0 +1,10 @@
+# Finetuning Evalhub
+
+This subcategory contains pipelines in the **Finetuning Evalhub** group:
+
+- [Lora Evalhub Pipeline](./lora/README.md): LoRA Training Pipeline with Eval Hub evaluation (KServe).
+- [Lora Minimal Evalhub Pipeline](./lora_minimal/README.md): LoRA Training Pipeline — Eval Hub Minimal.
+- [Osft Evalhub Pipeline](./osft/README.md): OSFT Training Pipeline with Eval Hub evaluation (KServe).
+- [Osft Minimal Evalhub Pipeline](./osft_minimal/README.md): OSFT Training Pipeline — Eval Hub Minimal.
+- [Sft Evalhub Pipeline](./sft/README.md): SFT Training Pipeline with Eval Hub evaluation (KServe).
+- [Sft Minimal Evalhub Pipeline](./sft_minimal/README.md): SFT Training Pipeline — Eval Hub Minimal.
