+++ pipelines/training/finetuning_evalhub/__init__.py
@@ -1,0 +1,5 @@
+"""Finetuning with Eval Hub Evaluation Pipelines.
+
+Pipelines that fine-tune models (LoRA, OSFT, SFT) and evaluate them
+via the Eval Hub service with KServe model serving.
+"""
