+++ pipelines/training/finetuning_evalhub/lora_minimal/metadata.yaml
