+++ pipelines/training/finetuning_evalhub/osft_minimal/metadata.yaml
