+++ pipelines/training/finetuning_evalhub/osft_minimal/pipeline.py
@@ -1,8 +1,9 @@
-"""SFT (Supervised Fine-Tuning) Training Pipeline — Eval Hub variant.
+"""OSFT (Orthogonal Subspace Fine-Tuning) Training Pipeline — Eval Hub Minimal variant.
 
-A 4-stage pipeline for standard supervised fine-tuning:
+A 4-stage pipeline for continual learning without catastrophic forgetting
+with minimal evaluation configuration:
 1. Dataset Download
-2. SFT Training (instructlab-training backend)
+2. OSFT Training (mini-trainer backend)
 3. Evaluation via Eval Hub (KServe InferenceService for model serving)
 4. Model Registry
 """
@@ -12,9 +13,11 @@ import kfp.kubernetes
 from kfp import dsl
 
 from components.data_processing.dataset_download import dataset_download
-from components.deployment.kubeflow_model_registry import kubeflow_model_registry
-from components.evaluation.evalhub_kserve import evalhub_evaluator_kserve
-from components.training.finetuning.sft import train_model
+from components.deployment.kubeflow_model_registry import (
+    kubeflow_model_registry as model_registry,
+)
+from components.evaluation.evalhub.kserve import evalhub_evaluator_kserve
+from components.training.finetuning.osft import train_model
 
 # =============================================================================
 # PVC Configuration (COMPILE-TIME settings)
@@ -22,13 +25,13 @@ from components.training.finetuning.sft import train_model
 PVC_SIZE = "50Gi"
 PVC_STORAGE_CLASS = "nfs-csi"
 PVC_ACCESS_MODES = ["ReadWriteMany"]
-PIPELINE_NAME = "sft-pipeline-evalhub"
+PIPELINE_NAME = "osft-pipeline-evalhub-easy"
 # =============================================================================
 
 
 @dsl.pipeline(
     name=PIPELINE_NAME,
-    description="SFT pipeline with Eval Hub evaluation via KServe, results optionally tracked in MLflow",
+    description="OSFT Eval Hub pipeline (Minimal): 5 leaderboard benchmarks, KServe serving, minimal config",
     pipeline_config=dsl.PipelineConfig(
         workspace=dsl.WorkspaceConfig(
             size=PVC_SIZE,
@@ -41,72 +44,46 @@ PIPELINE_NAME = "sft-pipeline-evalhub"
         ),
     ),
 )
-def sft_pipeline_evalhub(
+def osft_pipeline_evalhub_easy(
     # =========================================================================
     # KEY PARAMETERS (Required/Important) - Sorted by step
     # =========================================================================
     phase_01_dataset_man_data_uri: str,
     phase_02_train_man_train_batch: int = 128,
-    phase_02_train_man_epochs: int = 1,
-    phase_02_train_man_gpu: int = 1,
-    phase_02_train_man_model: str = "Qwen/Qwen2.5-1.5B-Instruct",
-    phase_02_train_man_tokens: int = 10000,
-    phase_02_train_man_workers: int = 4,
+    phase_02_train_man_train_epochs: int = 1,
+    phase_02_train_man_train_gpu: int = 1,
+    phase_02_train_man_train_model: str = "Qwen/Qwen2.5-1.5B-Instruct",
+    phase_02_train_man_train_tokens: int = 64000,
+    phase_02_train_man_train_unfreeze: float = 0.25,
+    phase_02_train_man_train_workers: int = 1,
     phase_03_eval_opt_evalhub_url: str = "",
-    phase_03_eval_opt_collection: str = "",
+    phase_03_eval_opt_mlflow_experiment: str = "",
+    phase_03_eval_opt_kserve_gpu_count: int = 1,
+    phase_03_eval_opt_kserve_cpu: str = "2",
+    phase_03_eval_opt_kserve_memory: str = "32Gi",
+    phase_03_eval_opt_timeout: int = 7200,
     phase_04_registry_man_address: str = "",
-    phase_04_registry_man_author: str = "pipeline",
-    phase_04_registry_man_name: str = "sft-model",
-    phase_04_registry_man_version: str = "1.0.0",
+    phase_04_registry_man_reg_author: str = "pipeline",
+    phase_04_registry_man_reg_name: str = "osft-model",
+    phase_04_registry_man_reg_version: str = "1.0.0",
     # =========================================================================
     # OPTIONAL PARAMETERS - Sorted by step
     # =========================================================================
     phase_01_dataset_opt_subset: int = 0,
-    phase_02_train_opt_annotations: str = "",
-    phase_02_train_opt_cpu: str = "4",
-    phase_02_train_opt_env_vars: str = (
-        "PYTORCH_CUDA_ALLOC_CONF=expandable_segments:True,NCCL_DEBUG=INFO,INSTRUCTLAB_NCCL_TIMEOUT_MS=600000"
-    ),
-    phase_02_train_opt_labels: str = "",
+    phase_02_train_opt_cpu: str = "8",
+    phase_02_train_opt_env_vars: str = "",
     phase_02_train_opt_learning_rate: float = 5e-6,
-    phase_02_train_opt_lr_warmup: int = 0,
-    phase_02_train_opt_lr_scheduler: str = "cosine",
     phase_02_train_opt_max_seq_len: int = 8192,
-    phase_02_train_opt_memory: str = "64Gi",
-    phase_02_train_opt_num_procs: str = "auto",
-    phase_02_train_opt_save_epoch: bool = True,
-    phase_02_train_opt_save_full_state: bool = False,
-    phase_02_train_opt_fsdp_sharding: str = "FULL_SHARD",
-    phase_02_train_opt_save_samples: int = 0,
-    phase_02_train_opt_seed: int = 42,
-    phase_02_train_opt_use_liger: bool = False,
+    phase_02_train_opt_memory: str = "32Gi",
+    phase_02_train_opt_use_liger: bool = True,
     phase_02_train_opt_runtime: str = "training-hub",
-    phase_03_eval_opt_benchmarks: list = [
-        {"id": "leaderboard_ifeval", "provider_id": "lm_evaluation_harness"},
-        {"id": "leaderboard_bbh", "provider_id": "lm_evaluation_harness"},
-        {"id": "leaderboard_mmlu_pro", "provider_id": "lm_evaluation_harness"},
-        {"id": "leaderboard_musr", "provider_id": "lm_evaluation_harness"},
-        {"id": "leaderboard_math_hard", "provider_id": "lm_evaluation_harness"},
-    ],
-    phase_03_eval_opt_mlflow_experiment: str = "",
-    phase_03_eval_opt_timeout: int = 7200,
-    phase_03_eval_opt_kserve_gpu_count: int = 1,
-    phase_03_eval_opt_kserve_cpu: str = "2",
-    phase_03_eval_opt_kserve_memory: str = "32Gi",
-    phase_04_registry_opt_description: str = "",
-    phase_04_registry_opt_format_name: str = "pytorch",
-    phase_04_registry_opt_format_version: str = "1.0",
     phase_04_registry_opt_port: int = 8080,
 ):
-    """SFT Training Pipeline with Eval Hub evaluation (KServe).
+    """OSFT Training Pipeline — Eval Hub Minimal.
 
-    A 4-stage ML pipeline for fine-tuning language models:
-
-    1) Dataset Download - Prepares training data from HuggingFace, S3, or HTTP
-    2) SFT Training - Fine-tunes using instructlab-training backend
-    3) Evaluation - Evaluates via Eval Hub with a KServe InferenceService for
-       model serving. Results optionally tracked in MLflow.
-    4) Model Registry - Registers trained model to Kubeflow Model Registry
+    Minimal-config variant that runs 5 leaderboard benchmarks (ifeval, bbh,
+    mmlu_pro, musr, math_hard) via KServe model serving — all public datasets,
+    no HF token required, no trust_remote_code issues.
 
     Prerequisites: Eval Hub and KServe must be installed on the cluster.
     The pipeline ServiceAccount needs RBAC permissions for
@@ -116,62 +93,37 @@ def sft_pipeline_evalhub(
     mount the model. The eval component uses the in-cluster ServiceAccount
     token for K8s API access.
 
-    Known limitations: Some HuggingFace datasets used by benchmarks require
-    trust_remote_code=True. The 5 default leaderboard benchmarks (ifeval,
-    bbh, mmlu_pro, musr, math_hard) work without it. For other benchmarks,
-    a custom provider ConfigMap with HF_DATASETS_TRUST_REMOTE_CODE=1 must
-    be configured. The base_model_name parameter is needed for tokenizer
-    resolution since the served model is a local checkpoint.
-
     Args:
         phase_01_dataset_man_data_uri: Dataset location (hf://, s3://, https://).
         phase_01_dataset_opt_subset: Limit to first N examples (0 = all).
-        phase_02_train_man_train_batch: Effective batch size per optimizer step.
-        phase_02_train_man_epochs: Number of training epochs.
-        phase_02_train_man_gpu: GPUs per worker.
-        phase_02_train_man_model: Base model (HuggingFace ID or path).
-        phase_02_train_man_tokens: Max tokens per GPU (memory cap).
-        phase_02_train_man_workers: Number of training pods.
-        phase_02_train_opt_annotations: Pod annotations (key=value,...).
+        phase_02_train_man_train_batch: Effective batch size (samples per optimizer step).
+        phase_02_train_man_train_epochs: Number of training epochs.
+        phase_02_train_man_train_gpu: GPUs per worker.
+        phase_02_train_man_train_model: Base model (HuggingFace ID or path).
+        phase_02_train_man_train_tokens: Max tokens per GPU (memory cap).
+        phase_02_train_man_train_unfreeze: Fraction to unfreeze
+            (0.1=minimal, 0.25=balanced, 0.5=strong).
+        phase_02_train_man_train_workers: Number of training pods.
         phase_02_train_opt_cpu: CPU cores per worker.
-        phase_02_train_opt_env_vars: Environment variables (KEY=VAL,...).
-        phase_02_train_opt_fsdp_sharding: FSDP strategy (FULL_SHARD,
-            HYBRID_SHARD).
-        phase_02_train_opt_labels: Pod labels (key=value,...).
-        phase_02_train_opt_learning_rate: Learning rate for training.
-        phase_02_train_opt_lr_scheduler: LR scheduler type (cosine, linear).
-        phase_02_train_opt_lr_warmup: Learning rate warmup steps.
-        phase_02_train_opt_max_seq_len: Maximum sequence length in tokens.
-        phase_02_train_opt_memory: Memory per worker (e.g., 64Gi).
-        phase_02_train_opt_num_procs: Processes per worker (auto or int).
+        phase_02_train_opt_env_vars: Env vars (KEY=VAL,...).
+        phase_02_train_opt_learning_rate: Learning rate. 5e-6 for OSFT.
+        phase_02_train_opt_max_seq_len: Max sequence length in tokens.
+        phase_02_train_opt_memory: RAM per worker.
         phase_02_train_opt_runtime: ClusterTrainingRuntime name.
-        phase_02_train_opt_save_epoch: Save checkpoint at each epoch.
-        phase_02_train_opt_save_full_state: Save full accelerate state.
-        phase_02_train_opt_save_samples: Number of samples to save (0 = none).
-        phase_02_train_opt_seed: Random seed for reproducibility.
         phase_02_train_opt_use_liger: Enable Liger kernel optimizations.
         phase_03_eval_opt_evalhub_url: Eval Hub API endpoint URL
             (empty = skip evaluation).
-        phase_03_eval_opt_collection: Eval Hub collection ID
-            (overrides benchmarks list). Available: "leaderboard-v2",
-            "safety-and-fairness-v1", "toxicity-and-ethical-principles".
-        phase_03_eval_opt_benchmarks: Benchmarks to evaluate. Defaults to 5
-            leaderboard benchmarks (ifeval, bbh, mmlu_pro, musr, math_hard)
-            that work without HF token or custom providers.
         phase_03_eval_opt_mlflow_experiment: MLflow experiment name
             (non-empty = enable, empty = disabled).
-        phase_03_eval_opt_timeout: Max seconds to wait for evaluation.
         phase_03_eval_opt_kserve_gpu_count: GPUs for the KServe predictor.
         phase_03_eval_opt_kserve_cpu: CPU for the KServe predictor.
         phase_03_eval_opt_kserve_memory: Pod memory for the KServe predictor.
+        phase_03_eval_opt_timeout: Max seconds to wait for evaluation.
         phase_04_registry_man_address: Model Registry address (empty = skip).
-        phase_04_registry_man_author: Author name for the registered model.
-        phase_04_registry_man_name: Model name in registry.
-        phase_04_registry_man_version: Semantic version.
-        phase_04_registry_opt_description: Model description for registry.
-        phase_04_registry_opt_format_name: Model format (pytorch, onnx).
-        phase_04_registry_opt_format_version: Model format version.
-        phase_04_registry_opt_port: Model Registry server port.
+        phase_04_registry_man_reg_author: Author name for the registered model.
+        phase_04_registry_man_reg_name: Model name in registry.
+        phase_04_registry_man_reg_version: Semantic version.
+        phase_04_registry_opt_port: Model registry server port.
     """
     # =========================================================================
     # Stage 1: Dataset Download
@@ -197,33 +149,30 @@ def sft_pipeline_evalhub(
     )
 
     # =========================================================================
-    # Stage 2: SFT Training
+    # Stage 2: OSFT Training (hardcoded safe defaults)
     # =========================================================================
     training_task = train_model(
         pvc_path=dsl.WORKSPACE_PATH_PLACEHOLDER,
         dataset=dataset_download_task.outputs["train_dataset"],
-        training_base_model=phase_02_train_man_model,
+        training_base_model=phase_02_train_man_train_model,
+        training_unfreeze_rank_ratio=phase_02_train_man_train_unfreeze,
         training_effective_batch_size=phase_02_train_man_train_batch,
-        training_max_tokens_per_gpu=phase_02_train_man_tokens,
+        training_max_tokens_per_gpu=phase_02_train_man_train_tokens,
         training_max_seq_len=phase_02_train_opt_max_seq_len,
         training_learning_rate=phase_02_train_opt_learning_rate,
-        training_seed=phase_02_train_opt_seed,
-        training_num_epochs=phase_02_train_man_epochs,
+        training_seed=42,
+        training_num_epochs=phase_02_train_man_train_epochs,
         training_use_liger=phase_02_train_opt_use_liger,
-        training_lr_scheduler=phase_02_train_opt_lr_scheduler,
-        training_lr_warmup_steps=phase_02_train_opt_lr_warmup,
-        training_checkpoint_at_epoch=phase_02_train_opt_save_epoch,
-        training_save_samples=phase_02_train_opt_save_samples,
-        training_accelerate_full_state_at_epoch=phase_02_train_opt_save_full_state,
-        training_fsdp_sharding_strategy=phase_02_train_opt_fsdp_sharding,
+        training_lr_scheduler="cosine",
+        training_lr_warmup_steps=0,
+        training_checkpoint_at_epoch=False,
+        training_save_final_checkpoint=True,
         training_envs=phase_02_train_opt_env_vars,
-        training_metadata_labels=phase_02_train_opt_labels,
-        training_metadata_annotations=phase_02_train_opt_annotations,
         training_resource_cpu_per_worker=phase_02_train_opt_cpu,
-        training_resource_gpu_per_worker=phase_02_train_man_gpu,
+        training_resource_gpu_per_worker=phase_02_train_man_train_gpu,
         training_resource_memory_per_worker=phase_02_train_opt_memory,
-        training_resource_num_procs_per_worker=phase_02_train_opt_num_procs,
-        training_resource_num_workers=phase_02_train_man_workers,
+        training_resource_num_procs_per_worker="auto",
+        training_resource_num_workers=phase_02_train_man_train_workers,
         training_runtime=phase_02_train_opt_runtime,
     )
     training_task.set_caching_options(False)
@@ -247,17 +196,22 @@ def sft_pipeline_evalhub(
     )
 
     # =========================================================================
-    # Stage 3: Evaluation via Eval Hub (KServe)
+    # Stage 3: Evaluation via Eval Hub (KServe, hardcoded safe defaults)
     # =========================================================================
     eval_task = evalhub_evaluator_kserve(
         pvc_mount_path=dsl.WORKSPACE_PATH_PLACEHOLDER,
         model_artifact=training_task.outputs["output_model"],
         evalhub_url=phase_03_eval_opt_evalhub_url,
-        collection_id=phase_03_eval_opt_collection,
-        benchmarks=phase_03_eval_opt_benchmarks,
+        benchmarks=[
+            {"id": "leaderboard_ifeval", "provider_id": "lm_evaluation_harness"},
+            {"id": "leaderboard_bbh", "provider_id": "lm_evaluation_harness"},
+            {"id": "leaderboard_mmlu_pro", "provider_id": "lm_evaluation_harness"},
+            {"id": "leaderboard_musr", "provider_id": "lm_evaluation_harness"},
+            {"id": "leaderboard_math_hard", "provider_id": "lm_evaluation_harness"},
+        ],
         evalhub_model_name="finetuned-model",
-        base_model_name=phase_02_train_man_model,
-        evalhub_job_name="sft-pipeline-eval",
+        base_model_name=phase_02_train_man_train_model,
+        evalhub_job_name="osft-pipeline-eval",
         evalhub_timeout=phase_03_eval_opt_timeout,
         evalhub_poll_interval=30,
         mlflow_experiment_name=phase_03_eval_opt_mlflow_experiment,
@@ -268,7 +222,7 @@ def sft_pipeline_evalhub(
     eval_task.set_caching_options(False)
     kfp.kubernetes.set_image_pull_policy(eval_task, "IfNotPresent")
 
-    for _task in [dataset_download_task, training_task, eval_task]:
+    for _task in [dataset_download_task, training_task]:
         kfp.kubernetes.use_secret_as_env(
             task=_task,
             secret_name="hf-token",
@@ -279,7 +233,7 @@ def sft_pipeline_evalhub(
     # =========================================================================
     # Stage 4: Model Registry
     # =========================================================================
-    model_registry_task = kubeflow_model_registry(
+    model_registry_task = model_registry(
         pvc_mount_path=dsl.WORKSPACE_PATH_PLACEHOLDER,
         input_model=training_task.outputs["output_model"],
         input_metrics=training_task.outputs["output_metrics"],
@@ -287,12 +241,12 @@ def sft_pipeline_evalhub(
         eval_results=eval_task.outputs["output_results"],
         registry_address=phase_04_registry_man_address,
         registry_port=phase_04_registry_opt_port,
-        model_name=phase_04_registry_man_name,
-        model_version=phase_04_registry_man_version,
-        model_format_name=phase_04_registry_opt_format_name,
-        model_format_version=phase_04_registry_opt_format_version,
-        model_description=phase_04_registry_opt_description,
-        author=phase_04_registry_man_author,
+        model_name=phase_04_registry_man_reg_name,
+        model_version=phase_04_registry_man_reg_version,
+        model_format_name="pytorch",
+        model_format_version="1.0",
+        model_description="",
+        author=phase_04_registry_man_reg_author,
         shared_log_file="pipeline_log.txt",
         source_pipeline_name=PIPELINE_NAME,
         source_pipeline_run_id=dsl.PIPELINE_JOB_ID_PLACEHOLDER,
@@ -305,6 +259,6 @@ def sft_pipeline_evalhub(
 
 if __name__ == "__main__":
     kfp.compiler.Compiler().compile(
-        pipeline_func=sft_pipeline_evalhub,
+        pipeline_func=osft_pipeline_evalhub_easy,
         package_path=__file__.replace(".py", ".yaml"),
     )
