+++ pipelines/training/finetuning_evalhub/sft/OWNERS
@@ -1,0 +1,6 @@
+approvers:
+  - briangallagher
+  - Fiona-Waters
+  - kramaranya
+  - MStokluska
+  - szaher
